+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/base/commands/UsersTransactions.kt
@@ -1,0 +1,5 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.commands
+
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.PostTransaction
+
+val usersTransactions = HashMap<Long, PostTransaction>()