+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/MediaGroupForwarder.kt
@@ -1,7 +1,7 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.forwarders
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
-import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeSync
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeBlocking
 import com.pengrad.telegrambot.TelegramBot
 import com.pengrad.telegrambot.model.Message
 import com.pengrad.telegrambot.model.request.*
@@ -16,7 +16,7 @@ class MediaGroupForwarder : Forwarder {
         return message.mediaGroupId != null
     }
 
-    override fun forward(bot: TelegramBot, targetChatId: Long, vararg messages: PostMessage): Map<PostMessage, Message> {
+    override suspend fun forward(bot: TelegramBot, targetChatId: Long, vararg messages: PostMessage): Map<PostMessage, Message> {
         val mediaGroups = mutableMapOf<String, MutableList<PostMessage>>()
         messages.forEach {
             postMessage ->
@@ -53,7 +53,7 @@ class MediaGroupForwarder : Forwarder {
             ) to it
         }.flatMap {
             (request, originals) ->
-            bot.executeSync(request).let {
+            bot.executeBlocking(request).let {
                 response ->
                 response.messages() ?.let {
                     (0 until originals.size).map {