+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/publishers/Publisher.kt
@@ -3,5 +3,5 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.publishers
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
 
 interface Publisher : Plugin {
-    fun publishPost(postId: Int)
+    suspend fun publishPost(postId: Int)
 }